add view, url for the all payments list
