+++ .env.example
@@ -8,4 +8,3 @@ POSTGRES_USER=postgres
 POSTGRES_PASSWORD=postgres
 DB_HOST=db
 DB_PORT=5432
-STRIPE_API_KEY=<your_api_key>
