+++ README.md
@@ -130,7 +130,6 @@ CLOUDPAYMENTS_PUBLIC_ID - public_id из личного кабинета Cloudpa
 
 CLOUDPAYMENTS_API_SECRET - API-ключ из личного кабинета Cloudpayments.
 
-STRIPE_API_KEY - API-ключ от Stripe.
 
 ### Cloudpayment
 Интернет-эйквайринг, позволяющий получать платежи онлайн с помощью банковской карты или различными методами  быстрой
@@ -151,41 +150,6 @@ https://cloudpayments.ru/integration
 - Взаимодействие осуществляется посредством сервиса Stripe, который агрегирует уже созданные платежи. Платежи проходят 
 через сервис Cloudpayments на стороне заказчика.
 - Реализована возможность проверки подключения к API Cloudpayments с помощью тест-кейса Django.
-
-### Stripe
-Сервис для приёма и обработки электронных платежей. 
-В проекте задействован как хранилище всего массива информации о платежах от двух эквайрингов.
-
-<details>
-    <summary><b>Подробнее о сервисе</b></summary>
-
-```shell
-https://docs.stripe.com/api
-```
-</details>
-
-Как реализовано: 
-
-- при создании объекта модели Cloudpayment и/или Mixplat срабатывает сигнал, 
-который запускает метод, создающий "Платежное намерение" в терминологии Stripe
-- в платежное намерение попадают не все данные о платеже, только обозначенные нами
-- чтобы клиент смог использовать функционал, необходима регистрация в Stripe и получение ключей доступа
-- все платежные намерения хранятся на стороне Stripe, при желании, их можно получить по API:
-
-```request
-GET /v1/payment_intents
-```
-
-```request
-stripe.PaymentIntent.list(limit=3)
-```
-<details>
-    <summary><b>Подробнее можно изучить здесь</b></summary>
-
-```shell
-https://docs.stripe.com/api/payment_intents/list
-```
-</details>
 
 
 ### Письма о статусе платежа
