+++ backend/requirements.txt
@@ -5,5 +5,4 @@ djangorestframework==3.14.0
 psycopg2-binary==2.9.3
 python-dotenv==1.0.1
 requests==2.31.0
-stripe==9.3.0
 whitenoise==6.6.0
